Rename Header to Heading; deprecate old one.
Heading seems like a more appropriate name, given that elements <h1>
through <h6> stand for "heading" (see HTMLHeadingElement), not "header"
(as I originally mistakenly thought). There is a <header> element in
HTML5, but that's usually a higher level container that contains a few
headings and a paragraph. This func returns a <h1>-<h6> element. Based
on that, heading seems like a more appropriate name.

> The HTML <h1>–<h6> elements represent six levels of section headings.
> <h1> is the highest section level and <h6> is the lowest.

> The HTML <header> element represents a group of introductory or
> navigational aids. It may contain some heading elements but also other
> elements like a logo, a search form, and so on.

Sources:

-	https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Heading_Elements
-	https://developer.mozilla.org/en-US/docs/Web/HTML/Element/header

The old deprecated Heading will be removed after 2 weeks.

Regenerate with latest version of vfsgen.