Add Go-unformatted alias for highlighting Go.
This is needed in contexts where markdownfmt is used to format markdown
files.

By default, markdownfmt applies gofmt formatting to fenced code blocks
with Go language. Most of the time, that matches the user's intention
of wanting to display properly formatted Go code. In rare cases, the
user intends to display purposefully unformatted Go code as is, while
still highlighting it as Go. This is what this Go-unformatted alias is
for.