Add minimal go.mod, go.sum.
The intention is to make it so that it's possible to import and use
this package from within Go modules (in addition to the well-supported
GOPATH workspace mode), and have it build without users having to
manually specify that github_flavored_markdown requires blackfriday v1.

Use the recently created blackfriday v1.5.2 tag.

Only specify the blackfriday version in go.mod and go.sum; others are
unspecified to avoid the added burden of maintaining and updating them.

Fixes #12.
Closes #20.